ai_agent-main/ai_agent-main/backend/requirements.txt
fastapi==0.109.0
langchain
langchain-community
langchain-google-genai
langgraph
chromadb
uvicorn==0.27.0
python-dotenv==1.0.1
pydantic
requests
beautifulsoup4
pdfplumber
google-search-results
langchain-openai
google-generativeai
langchain>=0.1.0
chromadb>=0.4.0
google-generativeai>=0.3.0
arxiv
pymupdf
scikit-learn>=1.2.0
sentence-transformers>=2.2.2
numpy